Add DefaultAlienWaveFactory
<interface, organisation, feature
